requirements.txt
python-dotenv==1.2.1
websockets==15.0.1
